requirements.txt
sounddevice
vosk
pyttsx3
PyYAML
numpy
tensorflow